# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LUMEN Knowledge Base API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/admin/users/{user_id}/spaces": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "List User Spaces Endpoint",
        "description": "查询用户已加入空间 + 可授予空间（API-054，REQ-050，维护态批5）。仅全局 admin。",
        "operationId": "list_user_spaces_endpoint_api_admin_users__user_id__spaces_get",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope_AdminUserSpacesView_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/spaces": {
      "get": {
        "tags": [
          "spaces"
        ],
        "summary": "List Spaces",
        "operationId": "list_spaces_api_spaces_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope_list_SpaceView__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/spaces/switch": {
      "post": {
        "tags": [
          "spaces"
        ],
        "summary": "Switch Space Endpoint",
        "operationId": "switch_space_endpoint_api_spaces_switch_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SwitchSpaceRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope_SwitchSpaceView_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/spaces/{space_id}/members": {
      "get": {
        "tags": [
          "space-members"
        ],
        "summary": "List Members Endpoint",
        "operationId": "list_members_endpoint_api_spaces__space_id__members_get",
        "parameters": [
          {
            "name": "space_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Space Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope_list_SpaceMemberView__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "space-members"
        ],
        "summary": "Add Member Endpoint",
        "operationId": "add_member_endpoint_api_spaces__space_id__members_post",
        "parameters": [
          {
            "name": "space_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Space Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddMemberRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope_SpaceMemberView_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/spaces/{space_id}/members/{user_id}": {
      "patch": {
        "tags": [
          "space-members"
        ],
        "summary": "Update Member Role Endpoint",
        "operationId": "update_member_role_endpoint_api_spaces__space_id__members__user_id__patch",
        "parameters": [
          {
            "name": "space_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Space Id"
            }
          },
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMemberRoleRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope_SpaceMemberView_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "space-members"
        ],
        "summary": "Remove Member Endpoint",
        "operationId": "remove_member_endpoint_api_spaces__space_id__members__user_id__delete",
        "parameters": [
          {
            "name": "space_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Space Id"
            }
          },
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope_NoneType_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/export/space": {
      "get": {
        "tags": [
          "export"
        ],
        "summary": "Export Space Endpoint",
        "operationId": "export_space_endpoint_api_export_space_get",
        "parameters": [
          {
            "name": "format",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "zip",
              "title": "Format"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AddMemberRequest": {
        "properties": {
          "email": {
            "type": "string",
            "title": "Email"
          },
          "role": {
            "type": "string",
            "title": "Role",
            "default": "member"
          }
        },
        "type": "object",
        "required": [
          "email"
        ],
        "title": "AddMemberRequest"
      },
      "AdminAvailableSpaceView": {
        "properties": {
          "space_id": {
            "type": "integer",
            "title": "Space Id"
          },
          "space_code": {
            "type": "string",
            "title": "Space Code"
          },
          "space_name": {
            "type": "string",
            "title": "Space Name"
          }
        },
        "type": "object",
        "required": [
          "space_id",
          "space_code",
          "space_name"
        ],
        "title": "AdminAvailableSpaceView"
      },
      "AdminJoinedSpaceView": {
        "properties": {
          "space_id": {
            "type": "integer",
            "title": "Space Id"
          },
          "space_code": {
            "type": "string",
            "title": "Space Code"
          },
          "space_name": {
            "type": "string",
            "title": "Space Name"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "joined_at": {
            "type": "string",
            "title": "Joined At"
          }
        },
        "type": "object",
        "required": [
          "space_id",
          "space_code",
          "space_name",
          "role",
          "joined_at"
        ],
        "title": "AdminJoinedSpaceView"
      },
      "AdminUserSpacesView": {
        "properties": {
          "joined": {
            "items": {
              "$ref": "#/components/schemas/AdminJoinedSpaceView"
            },
            "type": "array",
            "title": "Joined"
          },
          "available": {
            "items": {
              "$ref": "#/components/schemas/AdminAvailableSpaceView"
            },
            "type": "array",
            "title": "Available"
          }
        },
        "type": "object",
        "required": [
          "joined",
          "available"
        ],
        "title": "AdminUserSpacesView"
      },
...
      "ApiEnvelope_AdminUserSpacesView_": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "default": 0
          },
          "msg": {
            "type": "string",
            "title": "Msg",
            "default": "ok"
          },
          "data": {
            "$ref": "#/components/schemas/AdminUserSpacesView"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "ApiEnvelope[AdminUserSpacesView]"
      },
...
      "ApiEnvelope_NoneType_": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "default": 0
          },
          "msg": {
            "type": "string",
            "title": "Msg",
            "default": "ok"
          },
          "data": {
            "type": "null",
            "title": "Data"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "ApiEnvelope[NoneType]"
      },
...
      "ApiEnvelope_SpaceMemberView_": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "default": 0
          },
          "msg": {
            "type": "string",
            "title": "Msg",
            "default": "ok"
          },
          "data": {
            "$ref": "#/components/schemas/SpaceMemberView"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "ApiEnvelope[SpaceMemberView]"
      },
      "ApiEnvelope_SwitchSpaceView_": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "default": 0
          },
          "msg": {
            "type": "string",
            "title": "Msg",
            "default": "ok"
          },
          "data": {
            "$ref": "#/components/schemas/SwitchSpaceView"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "ApiEnvelope[SwitchSpaceView]"
      },
...
      "ApiEnvelope_list_SpaceMemberView__": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "default": 0
          },
          "msg": {
            "type": "string",
            "title": "Msg",
            "default": "ok"
          },
          "data": {
            "items": {
              "$ref": "#/components/schemas/SpaceMemberView"
            },
            "type": "array",
            "title": "Data"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "ApiEnvelope[list[SpaceMemberView]]"
      },
      "ApiEnvelope_list_SpaceView__": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "default": 0
          },
          "msg": {
            "type": "string",
            "title": "Msg",
            "default": "ok"
          },
          "data": {
            "items": {
              "$ref": "#/components/schemas/SpaceView"
            },
            "type": "array",
            "title": "Data"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "ApiEnvelope[list[SpaceView]]"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SpaceMemberView": {
        "properties": {
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "joined_at": {
            "type": "string",
            "title": "Joined At"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "name",
          "email",
          "role",
          "joined_at"
        ],
        "title": "SpaceMemberView"
      },
      "SpaceView": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "id",
          "code",
          "name"
        ],
        "title": "SpaceView"
      },
      "SwitchSpaceRequest": {
        "properties": {
          "space_id": {
            "type": "integer",
            "title": "Space Id"
          }
        },
        "type": "object",
        "required": [
          "space_id"
        ],
        "title": "SwitchSpaceRequest"
      },
      "SwitchSpaceView": {
        "properties": {
          "current_space_id": {
            "type": "integer",
            "title": "Current Space Id"
          }
        },
        "type": "object",
        "required": [
          "current_space_id"
        ],
        "title": "SwitchSpaceView"
      },
...
      "UpdateMemberRoleRequest": {
        "properties": {
          "role": {
            "type": "string",
            "title": "Role"
          }
        },
        "type": "object",
        "required": [
          "role"
        ],
        "title": "UpdateMemberRoleRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}